♻️ refactor: RightNav 타입스크립트 적용 및 Search 컴포넌트 적용 ##252 #260
